Make SpreeGlobalize optional

+++ db/migrate/20170827000002_add_layout_to_translatable_fields.rb
@@ -1,9 +1,13 @@
 class AddLayoutToTranslatableFields < SpreeExtension::Migration[4.2]
   def up
-    Spree::Page.add_translation_fields! layout: :string
+    if defined?(SpreeGlobalize)
+      Spree::Page.add_translation_fields! layout: :string
+    end
   end
 
   def down
-    remove_column :spree_page_translations, :layout
+    if defined?(SpreeGlobalize)
+      remove_column :spree_page_translations, :layout
+    end
   end
 end
